+++ packages/smart-contracts/typechain/factories/IEthRioStays__factory.ts
@@ -170,9 +170,9 @@ const _abi = [
       },
       {
         indexed: false,
-        internalType: "uint64",
+        internalType: "uint256",
         name: "pricePerNightWei",
-        type: "uint64",
+        type: "uint256",
       },
       {
         indexed: false,
@@ -213,9 +213,9 @@ const _abi = [
       },
       {
         indexed: false,
-        internalType: "uint64",
+        internalType: "uint256",
         name: "pricePerNightWei",
-        type: "uint64",
+        type: "uint256",
       },
       {
         indexed: false,
